add regex fomvalidation, add comments and refactoring

+++ public/js/utils.js
@@ -39,7 +39,7 @@ function addOnProductPage(product) {
     const selectColor = document.querySelector("#select-color");
 
     if (cart[product._id] == null) {
-
+        
         let teddySelected = {};
         teddySelected.id = product._id;
         teddySelected.name = product.name;
@@ -77,7 +77,11 @@ async function sendFormData(url, data) {
 // --------------- Formate les données à envoyer et envoie sur la page confirmation de commande --------------- //
 async function handleFormSubmit(e) {
     e.preventDefault();
-        
+
+    const formError = document.createElement('p');
+    formError.innerText = 'Veuillez remplir les champs indiqués';
+    formError.classList = 'error';
+
     const firstName = document.getElementById('first-name').value
     const lastName = document.getElementById('last-name').value
     const address = document.getElementById('address').value
@@ -85,34 +89,59 @@ async function handleFormSubmit(e) {
     const email = document.getElementById('email').value
     let products = [];
 
-    for(let product in getCart()) {
-        products.push(product);
+    const validName = /^(([a-zA-ZÀ-ÿ]+[\s\-]{1}[a-zA-ZÀ-ÿ]+)|([a-zA-ZÀ-ÿ]+))$/;
+    const validCity = /^(([a-zA-ZÀ-ÿ]+[\s\-]{1}[a-zA-ZÀ-ÿ]+)|([a-zA-ZÀ-ÿ]+)){1,10}$/;
+    const validMail = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9._-]{2,}\.[a-z]{2,4}$/;
+    const validAddress = /^(([a-zA-ZÀ-ÿ0-9]+[\s\-]{1}[a-zA-ZÀ-ÿ0-9]+)){1,10}$/;
+
+    console.log(validName.test(firstName.value));
+
+    // if(validName.test(firstName.value) !== true) {
+    //     firstName.parentNode.insertBefore(formError, firstName.nextSibling);
+    //     firstName.classList.add('error');
+    // }
+    // // if()
+
+    if ((validName.test(firstName) == true) &&
+        (validName.test(lastName) == true) &&
+        (validAddress.test(address) == true) &&
+        (validCity.test(city) == true) &&
+        (validMail.test(email) == true)) {
+
+            for(let product in getCart()) {
+                products.push(product);
+            }
+        
+            const contact = {
+                contact : {
+                    firstName,
+                    lastName,
+                    address,
+                    city,
+                    email
+                },
+                products
+            };
+        
+            const response = await sendFormData('/api/teddies/order', contact);
+        
+            if (response) {
+                //vide le localstorage
+                localStorage.clear();
+                // //crée nouveau localstorage pour stocker les données de commande
+                let orderValidation = {
+                    orderId: response.orderId,
+                    totalCart: document.getElementById("total-price").innerText,
+                    firstName: response.contact.firstName,
+                    email: response.contact.email
+                };
+                localStorage.setItem("orderValidation", JSON.stringify(orderValidation));
+                window.location = '/ordervalidation';
+            }
     }
-
-    const contact = {
-        contact : {
-            firstName,
-            lastName,
-            address,
-            city,
-            email
-        },
-        products
-    };
-
-    const response = await sendFormData('/api/teddies/order', contact);
-
-    if (response) {
-        //vide le localstorage
-        localStorage.clear();
-        // //crée nouveau localstorage pour stocker les données de commande
-        let orderValidation = {
-            orderId: response.orderId,
-            totalCart: document.getElementById("total-price").innerText,
-            firstName: response.contact.firstName,
-            email: response.contact.email
-        };
-        localStorage.setItem("orderValidation", JSON.stringify(orderValidation));
-        window.location = '/ordervalidation';
-    }
+    //  else {
+    //     alert("Merci de renseigner correctement les champs du formulaire afin de valider votre commande.");
+    //     return false;
+    // }
+    
 }
